[sim] fix signed/unsigned warning harder

+++ src/libraries/simulation_kernel/MIL_AgentServer.cpp
@@ -46,7 +46,7 @@ unsigned long FindMaxIdInFile( const tools::Path& filePath )
         boost::sregex_iterator it( line.begin(), line.end(), idRegex );
         const boost::sregex_iterator end;
         for( ; it != end; ++it )
-            for( size_t i = 1; i < it->size(); ++i )
+            for( int i = 1; i < static_cast< int >( it->size() ); ++i )
                 maxId = std::max( maxId,
                     boost::lexical_cast< unsigned long >( it->str( i ) ));
     }
